requirements.txt
django-dotenv==1.4.2
django-extensions==3.0.9
Django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6
Werkzeug==1.0.1
ipdb==0.13.4
ipython==7.18.1
pdbpp==0.10.2
django-model-utils==4.0.0
